Move tests to file
