+++ client/Player/PlayerView.test.tsx
@@ -8,7 +8,6 @@ import { Encounter } from "../Encounter/Encounter";
 import { env } from "../Environment";
 import { CurrentSettings, InitializeSettings } from "../Settings/Settings";
 import { buildEncounter } from "../test/buildEncounter";
-import { CombatFooter } from "./components/CombatFooter";
 import { PlayerView } from "./components/PlayerView";
 import { PlayerViewCombatant } from "./components/PlayerViewCombatant";
 import { PortraitModal } from "./components/PortraitModal";
@@ -142,60 +141,6 @@ describe("PlayerViewModel", () => {
     });
 
     expect(playerView.find(PortraitModal).length).toBe(0);
-  });
-});
-
-describe("Turn Timer", () => {
-  let encounter: Encounter;
-
-  beforeEach(() => {
-    InitializeSettings();
-    encounter = buildEncounter();
-  });
-
-  test("Player View round timer keeps time", () => {
-    jest.useFakeTimers();
-    encounter.AddCombatantFromStatBlock({
-      ...StatBlock.Default(),
-      HP: { Value: 10, Notes: "" },
-      Player: "player"
-    });
-    encounter.StartEncounter();
-    const playerViewState = encounter.GetPlayerView();
-
-    const combatFooter = Enzyme.shallow(
-      <CombatFooter
-        currentRound={playerViewState.RoundCounter}
-        timerVisible={true}
-        activeCombatantId={playerViewState.ActiveCombatantId}
-      />
-    );
-
-    jest.advanceTimersByTime(10000); // 10 seconds
-    expect(combatFooter.find(".turn-timer").text()).toBe("0:10");
-  });
-
-  test("Player View round timer stops when encounter stops", () => {
-    jest.useFakeTimers();
-    encounter.AddCombatantFromStatBlock({
-      ...StatBlock.Default(),
-      HP: { Value: 10, Notes: "" },
-      Player: "player"
-    });
-    encounter.StartEncounter();
-    encounter.EndEncounter();
-    const playerViewState = encounter.GetPlayerView();
-
-    const combatFooter = Enzyme.shallow(
-      <CombatFooter
-        currentRound={playerViewState.RoundCounter}
-        timerVisible={true}
-        activeCombatantId={playerViewState.ActiveCombatantId}
-      />
-    );
-
-    jest.advanceTimersByTime(10000); // 10 seconds
-    expect(combatFooter.find(".turn-timer").text()).toBe("0:00");
   });
 });
